Address PR #36 review comments (round 16)
- Abort traversal when eager -newer validation fails (matches GNU find
  which treats missing -newer refs as fatal argument errors)
- Use float64 comparison for -mmin +N/-N instead of truncated int64,
  fixing off-by-one-second boundary behavior

+++ interp/builtins/find/eval.go
@@ -181,9 +181,9 @@ func evalMmin(ec *evalContext, n int64, cmp int) bool {
 	diff := ec.now.Sub(modTime)
 	switch cmp {
 	case 1: // +N: strictly older than N minutes
-		return int64(diff.Seconds()) > n*60
+		return diff.Seconds() > float64(n*60)
 	case -1: // -N: strictly newer than N minutes
-		return int64(diff.Seconds()) < n*60
+		return diff.Seconds() < float64(n*60)
 	default: // N: ceiling-bucketed exact match
 		mins := int64(math.Ceil(diff.Minutes()))
 		return mins == n
